ARM: OMAP2+: edma: add support for suspend/resume
This patch adds suspend/resume support along with context save and
restore to edma module. All registers updated during probe, channel,
event allocation, param-set, interrupts are saved and restored.

This patch is tested on AM335x Beagle bone HS-MMC module.

TODO:
Currently all register are backed up and restored irrespective whether
there was any change from the values that was programmed or not.
This is likely to negatively impact the overall suspend/resume time.

Param-set saving can be optimized by looking at the actual usage. A
caching mechanism can be implemented and values can be cached when they
are changed. This way noting needs to be saved & restore is just a copy
from the cache.

Signed-off-by: Hebbar, Gururaja <gururaja.hebbar@ti.com>