ARM: OMAP2: Extend omap3_sram_restore_context to be generic.
The current SRAM context restore function is omap34xx specific, it re-pushes
the omap3_sram_configure_core_dpll code and the idle code. This is buggy
since omap_sram_push calls gen_pool_alloc, but there is no corresponding
gen_pool_free.

Instead of repushing the code on restore, save it on save, and restore it on
restore. This fixes the bugginess and allows us to re-use it on other SoCs.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>