ARM: AM33XX: Add RTC-only suspend
During RTC-only suspend, power is lost to the wkup domain, so we need to save
and restore the state of that domain. We also need to store some information
within the RTC registers so that u-boot can do the right thing at powerup.

The state is entering by getting the RTC to bring the pmic_power_en line low,
which will instruct the PMIC to disable the appropriate power rails after
putting DDR into self-refresh mode. To bring pmic_power_en low, we need to get
an ALARM2 event. Since we are running from SRAM at that point, it means
calculating what the next second is (via ASM) and programming that into the
RTC.

Signed-off-by: Steve Kipisz <s-kipisz2@ti.com>
Signed-off-by: Russ Dill <Russ.Dill@ti.com>