ARM: OMAP2+: Rework AM33XX IO PAD suspend patch to cover all of omap2
The previous patch, 'ARM: OMAP2+: AM33xx: PM: Add option to configure IO
PADs during suspend', allows the user to configure omap_mux modes during
suspend. The code is implemented in an am33xx specific way.

This conflicts with the RTC-only and hibernation patchsets that use a new generic
omax mux method to save and restore context. We modify the suspend IO PAD
patchset to use generic omap2 mux mechanisms. This reduces the complexity,
makes it easier to maintain, allows other omap2 platforms to take advantage of
the feature, and allows the RTC-only/hibernate patches to integrate with it.

The patch does not change the debugfs layout or functionality exept to allow
individual padconf changes to be disabled with a -1.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>