MFD: TPS65217: Add support for power supply notification
This patch adds support to notify/inform userspace on what power supplies
are connected to the TPS65217 (USB and/or AC). There is currently an issue
whereby MUSB ID polling causes a false positive for USB online. This is likely
due to leakage from ID to VUSB, likely through ESD protection diodes. The
consequence is that an a false ONLINE event occurs every 2s, followed 150ms
later by an OFFLINE event.

This is part of work to manage wakeup sources for hibernation and RTC-only
suspend.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>