ARM: OMAP2: Add functions to save and restore pinctrl context.
This adds a pair of context save/restore functions to save/restore the state
of a set of pinctrl registers. This simplifies some of the AM33XX PM code as
some of the pinctrl registers are lost when the per power domain loses power.
The pincrtl code can perform the necessary save/restore.

This will also be necessary for hibernation and RTC only sleep, as all
pinctrl registers all lost.

Signed-off-by: Russ Dill <Russ.Dill@gmail.com>