ARM: OMAP: Rework omap-rtc poweroff for faster poweroff.
The existing implementation could take up to 2 seconds to poweroff. The
new implementation takes up to a second.

This also fixes an issue with POWER_UP notification. Systems with
RTC power off control define the bit in the status register that
indicates that a power up has occurred with a bit that instead indicates
that ALARM2 (used for power off) has occurred.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>