PM / Hibernate: Add ability to store archdata in image.
In order for an external loader, such as u-boot, to restore a hibernation
image, additional architecture specific data, such as the resume address,
must be stored in the image.

This patch adds that ability by tacking on 1kB of space to the swsusp_info
header and calling out to swsusp_arch_add_info to fill that space.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>