Add clkdiv32k_ick to clk_24mhz_clkdm clockdomain
The debounce clock for GPIO1 depends is a part of clk_24mhz_clkdm clockdomain.
Not maintaining this association in DT can turn the domain off thus causing the
debounce clock to GPIO1 to turn off. This results in crash during boot when
gpio-keys driver is used.

We change gate-clock to ti,gate-clock for the clkdiv32k_ick as suggested by
Tero and make the association it to the clockdomain as done by Rajendra in his
earlier patch.

Additionally, the RTC entry in the device tree is incorrect. Although the IP can be
configured for either clk_32768_ck or clkdiv32k_ick the driver hardcodes it to
clkdiv32k_ick. The register to change this mux is contained within the RTC memory
space, so it would not be easy to encode into device tree. Just change it to the
clkdiv32k_ick for now.

Now that these clocks actually reference their clockdomain properly, the manual
proding of this clkdm can be removed from pm33xx.c.

Signed-off-by: Joel Fernandes <joelf@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Rajendra Nayak <rnayak@ti.com>
Signed-off-by: Russ Dill <Russ.Dill@ti.com>